requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
docutils==0.12
lockfile==0.10.2
psutil==3.1.1
python-daemon==2.0.5
setproctitle==1.1.9
tornado==4.2.1
wheel==0.24.0
